added different tabs
